Split TSX layout hotspots
