+++ src/app/features/code/code-workspace-defaults.ts
@@ -1,0 +1,33 @@
+import type { AppSettings } from '../../desktop/types'
+
+export const FALLBACK_APP_SETTINGS = {
+  chatModel: null,
+  chatThinkingLevel: null,
+  codeModel: null,
+  codeThinkingLevel: null,
+  gitCommitMessageModel: null,
+  gitCommitMessageThinkingLevel: 'off',
+  skillCreatorModel: null,
+  skillCreatorThinkingLevel: 'off',
+  composerStreamingBehavior: 'followUp',
+  dictationModelId: null,
+  dictationMaxDurationSeconds: 180,
+  showDictationButton: true,
+  favoriteFolders: [],
+  projectImportState: null,
+  preferredProjectLocation: null,
+  initializeGitOnProjectCreate: false,
+  gitOpsDefaultMode: 'commit',
+  gitDiffBaselineDefault: { kind: 'head' },
+  gitDiffRenderModeDefault: 'stacked',
+  gitDiffFileTreeDefaultVisible: true,
+  projectDeletionMode: 'pi-only',
+  useAgentsSkillsPaths: false,
+  howcodeNativeAskQuestions: false,
+  devUpdateBranch: false,
+  piTuiTakeover: false,
+  hoverToFocus: true,
+  hoverToBlur: false,
+  keybindings: {},
+  composerSendMode: 'enter',
+} satisfies AppSettings
